Move rated disability validation and view components
